fix: move PR B stamping to workflows/, keep no-growth files at baseline
Codex review found the previous commit (0f2c7e0) violated
scripts/check_architecture.py's ADR-061 no-growth debt ledger on three
files simultaneously:

- abicheck/cli_compare_release.py: 2000 lines > adoption baseline 1998
- abicheck/cli_compare_release_helpers.py: 1347 > baseline 1311
- tests/test_effective_config_digest.py: 1852 > baseline 1676

All three are debt.yaml-tracked legacy modules already at (or, for
cli_compare_release.py, just under) their frozen baseline -- new
behavior belongs in its ADR-061 target package, not appended to them.

- New abicheck/workflows/release_evaluation.py carries
  stamp_release_evaluation_config() (moved verbatim from
  cli_compare_release_helpers.py, which is back to its exact original
  1311 lines). Matches the root AGENTS.md ownership table: coordinating
  release-comparison behavior is a workflows/ responsibility. Untyped
  beyond Any to avoid pulling two more unclassified flat legacy modules
  (checker_types, pack_application) into workflows/'s permitted-import
  list purely for an annotation on a function that only does attribute
  access.
- cli_compare_release.py's own net diff is now 6 lines (one import, one
  call, one return) -- 1998 lines, exactly at baseline.
- The 8 new tests moved to a new tests/test_release_evaluation_config.py
  (self-contained, duplicating the three small fixture helpers rather
  than importing them from test_effective_config_digest.py, matching
  this suite's per-file convention). test_effective_config_digest.py is
  back to its exact original 1676 lines.

scripts/check_architecture.py: 0 errors. Full test suite for every
module importing the changed files (2006 tests) still passes.
